fix(sequential-donation): keep reset donation number till reach to id auto increment value

+++ includes/payments/class-give-sequential-donation-number.php
@@ -146,7 +146,9 @@ class Give_Sequential_Donation_Number {
 			get_option( '_give_reset_sequential_number' ) &&
 			( $number = give_get_option( 'sequential-ordering_number', 0 ) )
 		) {
-			delete_option( '_give_reset_sequential_number' );
+			if( Give()->sequential_donation_db->get_id_auto_increment_val() <= $number ){
+				delete_option( '_give_reset_sequential_number' );
+			}
 
 			return Give()->sequential_donation_db->insert( array(
 				'id'         => $number,
